requirements.txt
streamlit>=1.10.0
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.4.0
sympy>=1.8.0
pandas>=1.3.0
plotly>=5.3.0

# PDF Generation and Enhancement
reportlab>=3.6.0
PyPDF2>=2.0.0
pdfrw>=0.4.0
svglib>=1.0.0
qrcode>=7.0.0
Pillow>=9.0.0
weasyprint>=54.0
fpdf2>=2.5.0

# External dependencies (not installable via pip)
# LaTeX - Required for PDF generation
# Install TeX Live (Linux/macOS): https://tug.org/texlive/
# Install MiKTeX (Windows): https://miktex.org/
